PLAT-958 | Handle shortcuts in admin
